Moving ChannelListRow to component
